Changed nodesnumber type to size_t same as return function type

+++ singly_linked_lists/0-print_list.c
@@ -9,7 +9,7 @@
 
 size_t print_list(const list_t *h)
 {
-unsigned int nodesnumber;
+size_t nodesnumber;
 
 nodesnumber = 0;
